добавляет util.js
