+++ js/main.js
@@ -4,6 +4,3 @@ import './util.js';
 import './data.js';
 import { getRandomDescriptions } from './data.js';
 
-console.log(
-  getRandomDescriptions()
-)
